Nambah fitur

+++ admin/tambah-pesawat.php
@@ -5,6 +5,39 @@ include "../config/controller.php";
 
 $data_pesawat = select("SELECT * FROM pesawat");
 
+if(isset($_POST["tambah_pesawat"])){
+    if(tambah_pesawat($_POST, $_FILES)> 0){
+        echo "<script>
+        alert('Data pesawat Berhasil Ditambahkan');
+        document.location.href = 'tambah-pesawat.php';
+        </script>
+        ";      
+    }else{
+        echo "<script>
+        
+        alert('Data Pesawat Gagal Ditambahkan');
+        document.location.href = 'tambah-pesawat.php';
+        
+        </script>
+        ";  
+    }
+}
+
+if(isset($_POST["ubah_pesawat"])){
+    if(update_pesawat($_POST, $_FILES)>0){
+        echo "<script>
+        alert('Data Pesawat Berhasil Diubah');
+        document.location.href = 'tambah-pesawat.php';
+        </script>
+        ";
+    }else{
+        echo "<script>
+        alert('Data Pesawat Gagal Diubah');
+        document.location.href = 'tambah-pesawat.php';
+        </script>
+        ";
+    }
+}
 
 ?>
 
@@ -15,7 +48,7 @@ $data_pesawat = select("SELECT * FROM pesawat");
     <div class="d-flex flex-row-reverse">
         <!-- Button trigger modal -->
         <button type="button" class="btn btn-primary mb-5" data-bs-toggle="modal" data-bs-target="#tambah_pesawat">
-        Tambah Pesawat           
+        <i class="fa-solid fa-plus"></i> Tambah Pesawat           
         </button>
     </div>
     <table class="table table-hover table-info" id="table">
@@ -45,10 +78,72 @@ $data_pesawat = select("SELECT * FROM pesawat");
                 <td class="text-center"><?=$pesawat["kapasitas_penumpang"]?></td>
                 <td width="15%">
                     <button type="button" class="btn btn-success" data-bs-toggle="modal" data-bs-target="#edit_pesawat<?=$pesawat["id_pesawat"]?>">
-                            Ubah
+                    <i class="fa-regular fa-pen-to-square"></i>  Ubah
                     </button>
 
-                    <a href="hapus-pesawat.php" class="btn btn-danger">Hapus</a>
+                    <!-- Modal Edit Pesawat -->
+                    <div class="modal fade" id="edit_pesawat<?=$pesawat["id_pesawat"]?>" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+                          <div class="modal-dialog">
+                            <div class="modal-content">
+                              <div class="modal-header">
+                                <h1 class="modal-title fs-5" id="staticBackdropLabel">Ubah Data Teknisi</h1>
+                                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                              </div>
+                              <div class="modal-body">
+                                    <div class="container">
+                                    <form action="" method="post" autocomplete="off" enctype="multipart/form-data">
+                                        <div class="mt-3">
+                                            <input type="hidden" class="form-control" name="id_pesawat" placeholder="id Pesawat" required value="<?=$pesawat["id_pesawat"]?>">
+                                        </div>
+                                        <div class="mt-3">      
+                                            <label for="" class="form-label ">Nomor Registrasi <span style="color:red;">*</span></label>
+                                            <input type="number" class="form-control" name="nomor_registrasi" placeholder="Nomor registrasi" value="<?=$pesawat["no_registrasi"]?>" required>
+                                        </div>
+                                        <div class="mt-3">
+                                            <label for="" class="form-label">Nama Pesawat <span style="color:red;">*</span></label>
+                                            <input type="text" class="form-control" name="nama_pesawat" placeholder="Nama Pesawat" required value="<?=$pesawat["nama_pesawat"]?>" >
+                                        </div>
+                                        <div class="mt-3">
+                                            <label for="" class="form-label">Boeing Pesawat <span style="color:red;">*</span></label>
+                                            <input type="text" class="form-control" name="boeing_pesawat" placeholder="Boeing Pesawat" required value="<?=$pesawat["boieng_pesawat"]?>">
+                                        </div>
+                                        <div class="mt-3">
+                                            <label for="" class="form-label">Jenis Pesawat <span style="color:red;">*</span></label>
+                                            <select class="form-select form-select-sm" aria-label="Large select example" name="jenis_pesawat" required placeholder="Jenis Pesawat"
+                                            >
+                                                <option value="" disabled <?= empty($pesawat["jenis_pesawat"]) ? 'selected' : '' ?>>Jenis Pesawat</option>
+                                                <option value="Airbus" <?= $pesawat["jenis_pesawat"] == 'Airbus' ? 'selected' : '' ?>>Airbus</option>
+                                                <option value="Jet" <?= $pesawat["jenis_pesawat"] == 'Jet' ? 'selected' : '' ?>>Jet</option>
+                                                <option value="Boeing" <?= $pesawat["jenis_pesawat"] == 'Boeing' ? 'selected' : '' ?>>Boeing</option>
+                                            </select>
+                                        </div>
+
+                                        <div class="mt-3">      
+                                            <label for="" class="form-label ">Kapasitas Penumpang<span style="color:red;">*</span></label>
+                                            <input type="number" class="form-control" name="kapasitas_penumpang" placeholder="Nomor registrasi" required value="<?=$pesawat["kapasitas_penumpang"]?>">
+                                        </div>
+                                        <div class="mt-3">
+                                            <label for="" class="form-label">Gambar Pesawat Sebelumnya : </label>
+                                            <img src="../gambar_pesawat/<?=$pesawat["gambar_pesawat"]?>" alt="logo" height="40%" width="40%">
+                                        </div>
+                                        <div class="mt-3">
+                                            <label for="" class="form-label">Gambar Pesawat <span style="color:red;">*</span></label>
+                                            <input type="file" class="form-control" name="gambar_pesawat" placeholder="gambar Teknisi" required>
+                                        </div>
+                                        <div class="modal-footer">
+                                            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Tutup</button>
+                                            <button type="submit" class="btn btn-primary" name="ubah_pesawat">Ubah</button>
+                                        </div>
+
+                                    </div>
+
+                                    </form>
+                                </div>
+                            </div>
+                            </div>
+                          </div>
+                        </div>
+                    <a href="hapus-pesawat.php?id_pesawat=<?=$pesawat["id_pesawat"]?>" class="btn btn-danger" onclick="return confirm('Yakin mau menghapus pesawat ? ')"><i class="fa-regular fa-trash-can"></i> Hapus</a>
                 </td>
             </tr>
             <?php endforeach; ?>
@@ -76,7 +171,7 @@ $data_pesawat = select("SELECT * FROM pesawat");
       </div>
       <div class="modal-body">
             <div class="container" id="">
-            <form action="" method="post" autocomplete="off" enctype="multipart/form-data">
+            <form action="" method="post" autocomplete="off" enctype="multipart/form-data" id="myForm">
                 <div class="col">
                     <div class="mt-3">      
                         <label for="" class="form-label ">Nomor Registrasi <span style="color:red;">*</span></label>
@@ -91,27 +186,25 @@ $data_pesawat = select("SELECT * FROM pesawat");
                         <input type="text" class="form-control" name="boeing_pesawat" placeholder="Boeing Pesawat" required>
                     </div>
                     <div class="mt-3">
-                        <div class="btn-group">
-                            <button type="button" class="btn btn-danger dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false">
-                                Action
-                            </button>
-                            <ul class="dropdown-menu">
-                                <li><a class="dropdown-item" href="#">Action</a></li>
-                                <li><a class="dropdown-item" href="#">Another action</a></li>
-                                <li><a class="dropdown-item" href="#">Something else here</a></li>
-                                <li><hr class="dropdown-divider"></li>
-                                <li><a class="dropdown-item" href="#">Separated link</a></li>
-                            </ul>
-                        </div>
-
+                        <label for="" class="form-label">Jenis Pesawat <span style="color:red;">*</span></label>
+                        <select class="form-select form-select-sm" aria-label="Large select example" name="jenis_pesawat" required placeholder="Jenis Pesawat">
+                            <option disabled selected></option>
+                            <option value="Airbus">Airbus</option>
+                            <option value="Jet">Jet</option>
+                            <option value="Boeing">Boeing</option>
+                        </select>
+                    <div class="mt-3">      
+                        <label for="" class="form-label ">Kapasitas Penumpang<span style="color:red;">*</span></label>
+                        <input type="number" class="form-control" name="kapasitas_penumpang" placeholder="Nomor registrasi" required>
+                    </div>
                     </div>
                     <div class="mt-3">
-                        <label for="" class="form-label">Gambar Teknisi <span style="color:red;">*</span></label>
-                        <input type="file" class="form-control" name="gambar_teknisi" placeholder="gambar Teknisi" required>
+                        <label for="" class="form-label">Gambar Pesawat <span style="color:red;">*</span></label>
+                        <input type="file" class="form-control" name="gambar_pesawat" placeholder="gambar Teknisi" required>
                     </div>
                     <div class="modal-footer">
                         <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Tutup</button>
-                        <button type="submit" class="btn btn-primary" name="tambah_teknisi">Tambah</button>
+                        <button type="submit" class="btn btn-primary" name="tambah_pesawat">Tambah</button>
                     </div>
 
                 </div>
